refactor: 어드민 권한에 따라 분리했던 컨트롤러를 다시 병합
